fix-sdss-idr: 32-bit int
git-svn-id: svn+ssh://astrometry.net/svn/trunk/src/astrometry@12963 03a3cea6-2c03-0410-ac9b-9271b1c66c29

+++ util/fix_sdss_idr.py
@@ -45,7 +45,7 @@ def fix_sdss_idr(hdu):
 	newhdu._ffile = hdu._ffile
 	newhdu._datLoc = hdu._datLoc
 
-	newhdu.data = newhdu.data.astype(int)
+	newhdu.data = newhdu.data.astype(int32)
 	newhdu.data[newhdu.data < 0] += 2**16
 	print 'data type:', newhdu.data.dtype
 	print 'data range:', newhdu.data.min(), 'to', newhdu.data.max()
